+++ backend/src/main/java/com/bd2_team6/biteright/controllers/requests/create_requests/MealCreateRequest.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 public class MealCreateRequest {
-    private Integer mealTypeId;
+    private Long mealTypeId;
     private String name;
     private String description;
     private LocalDateTime mealDate;
