+++ backend/src/main/java/com/bd2_team6/biteright/controllers/requests/update_requests/MealContentUpdateRequest.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 public class MealContentUpdateRequest {
-    private Integer ingredientId;
-    private Integer mealId;
+    private Long ingredientId;
+    private Long mealId;
     private Integer ingredientAmount;
 }
