+++ backend/src/main/java/com/bd2_team6/biteright/entities/recipe_info/RecipeInfoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface RecipeInfoRepository extends JpaRepository<RecipeInfo, Integer> {
+public interface RecipeInfoRepository extends JpaRepository<RecipeInfo, Long> {
     Optional<RecipeInfo> findByRecipeName(String recipeName);
 }
