+++ backend/src/main/java/com/bd2_team6/biteright/entities/verification_code/VerificationCodeRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 
 @Repository
-public interface VerificationCodeRepository extends JpaRepository<VerificationCode, Integer>{
+public interface VerificationCodeRepository extends JpaRepository<VerificationCode, Long> {
     Optional<VerificationCode> findByUser(User user);
 }
